added mda support

+++ ims_stations_download.py
@@ -71,7 +71,8 @@ def parse_single_station(data):
               type=click.Path(exists=True), default=awd_path)
 @click.option('--axis_path', help='a full path to dem folder',
               type=click.Path(exists=True), default=axis_path)
-
+@click.option('--mda_path', help='a full path to where the ts-tm model files are',
+              type=click.Path(exists=True), default=work_yuval)
 def main_program(*args, **kwargs):
     from pathlib import Path
     window = kwargs['window']
@@ -79,14 +80,16 @@ def main_program(*args, **kwargs):
     gis_path = Path(kwargs['gis_path'])
     awd_path = Path(kwargs['awd_path'])
     axis_path = Path(kwargs['axis_path'])
-    ims_download(savepath, window)
-    ds = process_ims_stations(savepath, window, var='TD')
-    post_process_ims_stations(ds, window, savepath / 'TD', gis_path,
-                              awd_path, axis_path)
+    mda_path = Path(kwargs['mda_path'])
+    dsl = ims_download(savepath, window, save=False)
+    ds = process_ims_stations(savepath, window, var='TD', ds=dsl)
+    ds_axis = post_process_ims_stations(ds, window, savepath / 'TD', gis_path,
+                                        awd_path, axis_path)
+    produce_pw_all_stations(ds_axis, axis_path, mda_path)
     return
 
 
-def ims_download(savepath, window):
+def ims_download(savepath, window, save=False):
     """
     Downloads a 10mins parameter from the IMS for the last <window> hours for all stations.
 
@@ -119,6 +122,7 @@ def ims_download(savepath, window):
     # use only active 10mins stations:
     stations_df = stations_df[stations_df['timebase'] == 10]
     stations_df = stations_df[stations_df['active']]
+    ds_list = []
     for i, row in stations_df.iterrows():
         st_id = row['stationId']
         st_name = row['name']
@@ -149,12 +153,15 @@ def ims_download(savepath, window):
         ds.attrs['lat'] = lat
         ds.attrs['lon'] = lon
         ds.attrs['station_id'] = st_id
-        filename = '{}.nc'.format('-'.join(st_name.split(' ')))
-        save_ncfile(ds, savepath, filename)
-    return
+        if save:
+            filename = '{}.nc'.format('-'.join(st_name.split(' ')))
+            save_ncfile(ds, savepath, filename)
+        else:
+            ds_list.append(ds)
+    return ds_list
 
 
-def process_ims_stations(mainpath, window, var='TD'):
+def process_ims_stations(mainpath, window, var='TD', ds=None):
     import os
     import xarray as xr
     from aux_gps import path_glob
@@ -168,9 +175,11 @@ def process_ims_stations(mainpath, window, var='TD'):
         logger.info('created {}.'.format(savepath))
     else:
         logger.info('{} already exist.'.format(savepath))
-
-    files = path_glob(mainpath, '*.nc')
-    dsl = [xr.load_dataset(x) for x in files]
+    if ds is None:
+        files = path_glob(mainpath, '*.nc')
+        dsl = [xr.load_dataset(x) for x in files]
+    else:
+        dsl = ds
     ds_list = []
     for ds in dsl:
         try:
@@ -193,7 +202,8 @@ def process_ims_stations(mainpath, window, var='TD'):
     filename = 'IMS_{}_{}-{}.nc'.format(var, st_str, end_str)
     save_ncfile(ds, savepath, filename)
     # finally delete all nc files:
-    [x.unlink() for x in files]
+    if ds is None:
+        [x.unlink() for x in files]
     return ds
 
 
@@ -212,8 +222,8 @@ def post_process_ims_stations(ds, window, savepath, gis_path, dem_path,
     ds = analyse_10mins_ims_field(ds=ds, var='TD', gis_path=gis_path,
                                   dem_path=dem_path)
     ds_axis = IMS_interpolating_to_GNSS_stations_israel(
-            dt=None, start_year=str(now_dt.year), verbose=True, savepath=None,
-            network='axis', ds_td=ds, cut_days_ago=None, axis_path=axis_path)
+        dt=None, start_year=str(now_dt.year), verbose=True, savepath=None,
+        network='axis', ds_td=ds, cut_days_ago=None, axis_path=axis_path)
     now_dt = pd.Timestamp.utcnow().floor('H')
     names = produce_rinex_filenames_at_time_window(end_dt=now_dt,
                                                    window=window)
@@ -221,7 +231,43 @@ def post_process_ims_stations(ds, window, savepath, gis_path, dem_path,
     end_str = names[-1][4:8]
     filename = 'AXIS_TD_{}-{}.nc'.format(st_str, end_str)
     save_ncfile(ds_axis, savepath, filename)
-    return
+    return ds_axis
+
+
+def produce_pw_all_stations(ds, axis_path, mda_path):
+    from PW_stations import load_mda
+    from PW_stations import produce_GNSS_station_PW
+    from aux_gps import fill_na_xarray_time_series_with_its_group
+    from aux_gps import path_glob
+    from aux_gps import save_ncfile
+    import xarray as xr
+    # first load mda:
+    mda = load_mda(mda_path)
+    # now loop over each station, produce pwv and save:
+    st_dirs = path_glob(axis_path, '*/')
+    st_dirs = [x for x in st_dirs if x.is_dir()]
+    st_dirs = [x for x in st_dirs if not x.as_posix().split('/')[-1].isnumeric()]
+    assert len(st_dirs) == 27
+    for st_dir in st_dirs:
+        station = st_dir.as_posix().split('/')[-1]
+        last_file = sorted(path_glob(st_dir/'dr/ultra', '*.nc'))[-1]
+        last_file_str = last_file.as_posix().split('/')[-1][4:13]
+        wet = xr.load_dataset(last_file)['WetZ']
+        wet_error = xr.load_dataset(last_file)['WetZ_error']
+        wet.name = station
+        wet_error = station
+        # resample temp to 5 mins and reindex to wet delay time:
+        t = ds[station].resample(time='5T').ffill().reindex_like(wet.time)
+        # fill in NaNs with mean hourly signal:
+        t_new = fill_na_xarray_time_series_with_its_group(t, grp='hour')
+        pwv = produce_GNSS_station_PW(wet, t_new, mda=mda,
+                                      model_name='LR', plot=False)
+        pwv_error = produce_GNSS_station_PW(wet_error, t_new, mda=mda,
+                                            model_name='LR', plot=False)
+        pwv_error.name = '{}_error'.format(pwv.name)
+        pwv_ds = xr.merge([pwv, pwv_error])
+        filename = '{}{}_PWV.nc'.format(station, last_file_str)
+        save_ncfile(pwv_ds, st_dir/'dr/ultra', filename)
 
 # def check_ds_last_datetime(ds, fmt=None):
 #     """return the last datetime of the ds"""
@@ -245,4 +291,3 @@ def post_process_ims_stations(ds, window, savepath, gis_path, dem_path,
 
 if __name__ == '__main__':
     main_program()
-
